+++ webview-ui/src/panels/Welcome.svelte
@@ -78,6 +78,11 @@
                 There was an issue in tmc-langs that caused the extension to fail to detect the
                 project directory within a submission after downloading it.
             </p>
+            <h4>Fixed error messages not displaying in some cases when running tests</h4>
+            <p>
+                Certain types of errors were not being displayed when they occurred while running
+                tests.
+            </p>
             <h3>3.0.6 - 2024-11-07</h3>
             <h4>Added a warning for a large number of open exercises</h4>
             <p>